Remove python local hack for line numbers
This headache was caused by python-mode as it was insisting on line
numbers. It also caused other woes, namely with code checking. It would
run pylint on all imported files as well causing performance issues and
popping up a quickfix window which annoyed the hell out of me. This is
why python-mode was removed in previous commit.